feat: implement inngest function "Recall Assistant"
Move inngest functions to own directory.
Move common function functionality to utility class.
