+++ .env.example
@@ -12,6 +12,7 @@ NEXT_PUBLIC_STREAM_VIDEO_API_KEY=
 STREAM_VIDEO_API_SECRET=
 
 OPENAI_API_KEY=
+GEMINI_API_KEY=
 
 POLAR_ACCESS_TOKEN=
 POLAR_SERVER=
