+++ src/app/api/inngest/route.ts
@@ -1,7 +1,7 @@
 import { serve } from 'inngest/next';
 
 import { inngest } from '@/inngest/client';
-import { meetingsProcessing } from '@/inngest/functions';
+import { meetingsProcessing } from '@/inngest/functions/function-summarizer';
 
 export const { GET, POST, PUT } = serve({
   client: inngest,
